requirements.txt
# AWS Bedrock (TwelveLabs Pegasus + Marengo + Claude Haiku) and S3
boto3>=1.34.0

# Numerical + embedding similarity
numpy>=1.26.0
scikit-learn>=1.4.0

# Tabular handling + CSV export
pandas>=2.2.0

# Geocoding via Overture Maps — query public S3 parquet directly
duckdb>=1.0.0

# Map generation
folium>=0.16.0

# Streamlit frontend
streamlit>=1.34.0

# GeoJSON export
geojson>=3.1.0

# Config loading
PyYAML>=6.0.0
python-dotenv>=1.0.0

# Notes:
# - DOCX text extraction uses stdlib zipfile + regex — no python-docx dep.
# - Claude is called via Bedrock (boto3) — no anthropic SDK dep.
# - Geocoding uses Overture Maps via DuckDB — no geopy/Nominatim dep.
# - PDF support not yet wired (pdfplumber removed); add back if needed.
